cocosgame/requirements.txt
cocos2d==0.6.10
numpy==2.3.1
pygame==2.6.1
pyglet==1.5.31
six==1.17.0
